gondola-core/rust/gondola-core/pyproject.toml
[project]
name            = 'gondola'
#version         = "0.11.1"
dynamic         = ['version']
description     = "GAPS Online and Offline Analysis software"
authors         = [{ name = "J.A.Stoessl, G.Tytus", email = "stoessl@hawaii.edu" }]
readme          = "README.md"
license         = { text = "GPLV3" }
requires-python = ">=3.11"
dependencies = [
    "tqdm>=4.0",    
    "numpy>=2.3",
    "matplotlib>=3.10",
    "charmingbeauty>=0.1",
    "pydashi>=2",
    "cmocean>=4",
    "loguru>=0.7",
    "vtk>=9",
    "iminuit>=2",
    "polars>=1.33"
]

[build-system]
requires      = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[tool.maturin]
# "extension-module" tells pyo3 we want to build an extension module (skips linking against libpython.so)
#features     = ["pyo3/extension-module"]
#all-features  = true
features      = ["random","database","advanced-algorithms","pybindings","tof-liftof"]
python-source = "python"
module-name   = "gondola.gondola_core"



